Separate Settings,Input
