+++ Assets/Scripts/Input/InputManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 114f8a5e4c8df6844a866ba5efc32a1f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
