+++ Assets/Scripts/Setting.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: eff6a0523c75d4a4db5a3c61d075d6d7
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
