+++ Assets/Scripts/Setting/Settings.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c5b50cfeabd903246b2cc7b96c13b3c7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
